feat: Split sso loaders
